move hs_matrix_k.h and .cpp to source_hamilt
