+++ source/source_lcao/module_operator_lcao/operator_lcao.h
@@ -4,7 +4,7 @@
 #include "source_hamilt/matrixblock.h"
 #include "source_hamilt/operator.h"
 #include "source_hamilt/module_hcontainer/hcontainer.h"
-#include "source_lcao/hs_matrix_k.h"
+#include "source_hamilt/hs_matrix_k.h"
 
 namespace hamilt {
